requirements.txt
crewai
crewai_tools
python-dotenv
langchain_openai
langchain_community
langchain_huggingface